stock_models/requirements.txt
absl-py==2.1.0
atari-py==0.2.9
autopep8==2.3.2
beautifulsoup4==4.12.3
blinker==1.9.0
bs4==0.0.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.0
comtypes==1.4.8
contourpy==1.3.1
cssselect==1.2.0
cycler==0.12.1
easytrader==0.23.0
easyutils==0.1.7
empyrical==0.5.5
filelock==3.16.1
Flask==3.1.0
fonttools==4.54.1
fsspec==2024.10.0
futu-api==8.7.4708
grpcio==1.67.1
gym==0.26.2
gym-notices==0.0.8
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
lxml==5.3.0
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.10.0.84
pandas==2.2.3
pandas-datareader==0.10.0
pillow==11.0.0
pluggy==1.5.0
protobuf==3.20.3
pycodestyle==2.12.1
pycryptodome==3.21.0
pyfolio==0.9.2
pyparsing==3.2.0
pyquery==2.0.1
pytest==8.3.3
pytz==2024.2
pywinauto==0.6.6
requests==2.32.3
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
simplejson==3.19.3
soupsieve==2.6
stable-baselines3==1.2.0
stockstats==0.6.2
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
tomli==2.1.0
torch==2.5.1
tqdm==4.67.0
triton==3.1.0
tushare==1.4.13
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==3.1.3
